refactor ai lib module
