+++ assets/js/lib/ai/index.js
@@ -1,29 +1,6 @@
 // SPDX-FileCopyrightText: SUSE LLC
 // SPDX-License-Identifier: Apache-2.0
 
-import { get } from 'lodash';
-
-import GeminiIcon from '@static/gemini-logo.svg';
-import OpenAIIcon from '@static/openai-logo.svg';
-import AnthropicIcon from '@static/anthropic-logo.svg';
-
-const providerRenderingConfig = {
-  googleai: {
-    icon: GeminiIcon,
-    label: 'Google Gemini',
-  },
-  openai: {
-    icon: OpenAIIcon,
-    label: 'OpenAI GPT',
-  },
-  anthropic: {
-    icon: AnthropicIcon,
-    label: 'Anthropic Claude',
-  },
-};
-
-export const getProviderLabel = (provider) =>
-  get(providerRenderingConfig, [provider, 'label'], provider);
-
-export const getProviderIcon = (provider) =>
-  get(providerRenderingConfig, [provider, 'icon']);
+export { getProviderLabel, getProviderIcon } from './providers';
+export { WebSocketAIAgent, extractMessageText } from './WebSocketAIAgent';
+export { CONNECTION_STATUS } from './connectionStatus';
