+++ assets/js/lib/ai/providers.test.js
@@ -1,9 +1,9 @@
 // SPDX-FileCopyrightText: SUSE LLC
 // SPDX-License-Identifier: Apache-2.0
 
-import { getProviderLabel, getProviderIcon } from '.';
+import { getProviderLabel, getProviderIcon } from './providers';
 
-describe('aizz', () => {
+describe('providers', () => {
   describe('getProviderLabel', () => {
     it('should return the provider label', () => {
       expect(getProviderLabel('googleai')).toBe('Google Gemini');
